Teachable_Moments/requirements.txt
# Core ML/DL
torch>=2.0.0
transformers>=4.30.0
peft>=0.5.0
accelerate>=0.20.0
datasets>=2.14.0

# Reinforcement Learning
gymnasium>=0.29.0
minigrid>=2.3.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0

# API Clients (for teacher models)
openai>=1.0.0

# Experiment Tracking (optional)
wandb>=0.15.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
ruff>=0.0.280
mypy>=1.4.0

# Type stubs
types-PyYAML>=6.0.0
types-tqdm>=4.65.0
